// src/web/src/api/generated/model/importAnalysisDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ImportAnalysisDtoSuggestedMapping } from './importAnalysisDtoSuggestedMapping';

export interface ImportAnalysisDto {
  columns: string[];
  sampleRows: string[][];
  suggestedMapping: ImportAnalysisDtoSuggestedMapping;
  rowCount: number;
}