// src/web/src/api/generated/model/importPreviewDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ImportIssueDto } from './importIssueDto';
import type { ImportPreviewItemDto } from './importPreviewItemDto';

export interface ImportPreviewDto {
  readyCount: number;
  errorCount: number;
  skippedCount: number;
  issues: ImportIssueDto[];
  sample: ImportPreviewItemDto[];
}