// src/web/src/api/generated/model/previewItemImportBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IFormFile } from './iFormFile';
import type { Unit } from './unit';

export type PreviewItemImportBody = {
  NameColumn?: number;
  SkuColumn?: number;
  BarcodeColumn?: number;
  UnitColumn?: number;
  CategoryColumn?: number;
  GroupNameColumn?: number;
  PurchasePriceColumn?: number;
  SalePriceColumn?: number;
  MinStockColumn?: number;
  InitialStockColumn?: number;
  DefaultUnit?: Unit;
  File?: IFormFile;
};